rename pir/transforms/fusion to pir/transforms/gpu (#62759)
* rename pir/transforms/fusion to pir/transforms/gpu

* fix

* fix
