+++ paddle/fluid/pir/transforms/general/dead_code_elimination_pass.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/pir/transforms/dead_code_elimination_pass.h"
+#include "paddle/fluid/pir/transforms/general/dead_code_elimination_pass.h"
 #include <cstdint>
 
 #include "paddle/fluid/pir/dialect/operator/ir/op_dialect.h"
