+++ paddle/fluid/pir/transforms/general/map_op_to_another_pass.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/pir/transforms/map_op_to_another_pass.h"
+#include "paddle/fluid/pir/transforms/general/map_op_to_another_pass.h"
 
 #include "paddle/fluid/pir/dialect/operator/ir/pd_op.h"
 #include "paddle/fluid/pir/drr/include/drr_pattern_base.h"
