+++ paddle/fluid/pir/transforms/gpu/fused_dot_product_attention_pass.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/pir/transforms/fusion/fused_dot_product_attention_pass.h"
+#include "paddle/fluid/pir/transforms/gpu/fused_dot_product_attention_pass.h"
 
 #include "paddle/fluid/pir/dialect/operator/ir/pd_op.h"
 #include "paddle/fluid/pir/drr/include/drr_pattern_base.h"
